Move aggregate and aggregate_finalized to aggregator
