+++ query/src/pipelines/transforms/aggregator/mod.rs
@@ -1,3 +1,6 @@
 mod aggregator_u32;
+mod aggregator;
 
+pub use aggregator::aggregate;
+pub use aggregator::aggregate_finalized;
 // pub use aggregator_u32::AggregatorWithU32;
